+++ sickrage/core/logger.py
@@ -21,10 +21,12 @@ from __future__ import unicode_literals
 import logging
 import os
 import pkgutil
+import platform
 import re
 from logging import FileHandler, CRITICAL, DEBUG, ERROR, INFO, WARNING
 from logging.handlers import RotatingFileHandler
 
+import raven
 from raven.handlers.logging import SentryHandler
 from unidecode import unidecode
 
@@ -86,7 +88,13 @@ class Logger(logging.getLoggerClass()):
         self.handlers = []
 
         # sentry log handler
-        sentry_handler = SentryHandler(client=sickrage.app.sentry_client)
+        sentry_client = raven.Client('https://aedcce983da74dbb8a74384cb31fe56b:9642f0779bfa43bf8b849253d7c68972'
+                                     '@sentry.sickrage.ca/3?verify_ssl=0')
+
+        sentry_tags = {'app_version': sickrage.version(),
+                       'platform': platform.platform()}
+
+        sentry_handler = SentryHandler(client=sentry_client, tags=sentry_tags)
         sentry_handler.setLevel(self.logLevels['ERROR'])
         self.addHandler(sentry_handler)